src/dynastore/modules/spanner/requirements.txt
google-cloud-spanner>=3.14.1,<4.0.0